refactor(store-core): extract shared handler helpers into handlers/_shared.ts (audit P2-3)
Enabling slice for the P2-3 split. Move the cross-cutting handler helpers
out of the barrel so sibling family modules can import them without a
circular dependency back on ./index:

- parseStringField / parseRawStringField / parseEnumField (field parsers)
- resolveSessionId (session-resolution primitive, 21 call sites)
- SessionPatch (the session-scoped patch type)

Pure move, no logic change: the moved bodies are byte-identical to the
removed lines (proven by diff, modulo the added `export` keyword on the
three parsers, which were module-private). index.ts now imports them from
./_shared and re-exports resolveSessionId + SessionPatch (the only two that
were part of the barrel's public surface); the three field parsers stay
private (not re-exported), so the public surface is unchanged.

This unblocks the remaining stateful families (budget, plan, checkpoint,
agent, stream/tool, …), all of which depend on these helpers.

handlers/index.ts: 5209 -> 5150 lines.
